+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -156,7 +156,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-05-29T10:37:53.086Z",
+    "2026-05-29T11:58:31.816Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {