+++ packages/orq-rc/RELEASES.md
@@ -9569,3 +9569,13 @@ Based on:
 - [typescript v4.10.0-rc.55] packages/orq-rc
 ### Releases
 - [NPM v4.10.0-rc.55] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.55 - packages/orq-rc
+
+## 2026-05-29 12:01:40
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.768.2 (2.889.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.10.0-rc.56] packages/orq-rc
+### Releases
+- [NPM v4.10.0-rc.56] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.56 - packages/orq-rc